utils/notebook_requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autopep8==2.3.2
babel==2.17.0
beautifulsoup4==4.13.4
black==25.1.0
bleach==6.2.0
celery==5.4.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.1
flower==2.0.1
fqdn==1.5.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbqa==1.9.1
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==2.2.6
openpyxl==3.1.5
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.26.0
pytest==8.4.0
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
redis==5.3.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
Send2Trash==1.8.3
setuptools==80.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
terminado==0.18.1
tinycss2==1.4.0
tokenize_rt==6.2.0
tomli==2.2.1
tornado==6.5.1
traitlets==5.14.3
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
